validaciones procesos y periodos

+++ backend/src/controllers/proceso.controller.js
@@ -1,5 +1,9 @@
 "use strict";
 
+const {
+  procesoIdSchema,
+  finalizadoSchema,
+} = require("../schema/proceso.schema");
 const { respondSuccess, respondError } = require("../utils/resHandler");
 const { handleError } = require("../utils/errorHandler");
 const ProcesoService = require("../services/proceso.service");
@@ -21,6 +25,7 @@ async function getProcesos(req, res) {
 async function createProceso(req, res) {
   try {
     const { body } = req;
+
     const [newProceso, procesoError] = await ProcesoService.createProceso(body);
 
     if (procesoError) return respondError(req, res, 400, procesoError);
@@ -35,9 +40,34 @@ async function createProceso(req, res) {
   }
 }
 
+async function updateFinalizadoProceso(req, res) {
+  try {
+    const { params, body } = req;
+    const { error: paramsError } = procesoIdSchema.validate(params);
+    if (paramsError) return respondError(req, res, 400, paramsError.message);
+
+    const { error: bodyError } = finalizadoSchema.validate(body);
+    if (bodyError) return respondError(req, res, 400, bodyError.message);
+
+    const [proceso, procesoError] =
+      await ProcesoService.updateFinalizadoProceso(params.id, body.finalizado);
+    if (procesoError) return respondError(req, res, 400, procesoError);
+    if (!proceso) {
+      return respondError(req, res, 400, "No se actualizó el proceso");
+    }
+
+    respondSuccess(req, res, 200, proceso);
+  } catch (error) {
+    handleError(error, "proceso.controller -> updateFinalizadoProceso");
+    respondError(req, res, 500, "No se actualizó el proceso");
+  }
+}
+
 async function deleteProceso(req, res) {
   try {
     const { params } = req;
+    const { error: paramsError } = procesoIdSchema.validate(params);
+    if (paramsError) return respondError(req, res, 400, paramsError.message);
 
     const [proceso, procesoError] = await ProcesoService.deleteProceso(
       params.id,
@@ -57,5 +87,6 @@ async function deleteProceso(req, res) {
 module.exports = {
   getProcesos,
   createProceso,
+  updateFinalizadoProceso,
   deleteProceso,
 };
